App builder

See merge request root/oxzion3.0!26

+++ api/v1/lib/Oxzion/src/Utils/FileUtils.php
@@ -106,8 +106,10 @@ class FileUtils
                 self::deleteDirectoryContents( $dir.'/'.$file );
             }
             rmdir( $dir );
-        } else {
+        } else if(file_exists($dir)){
             unlink( $dir );
+        } else {
+            return;
         }
     }
     public static function getFiles($directory)
